+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.7.25 (Upcoming release)
 
+- Move Marcos's original normalization to a seperate function. Now Marcos can take any arbitrary normalization method as an optional argument.
+
 
 ### 0.7.24